+++ ai_animation/src/types/map.ts
@@ -8,6 +8,7 @@ export enum ProvTypeENUM {
 }
 
 export enum PowerENUM {
+  ENGLAND = "England",
   FRANCE = "France",
   TURKEY = "Turkey",
   AUSTRIA = "Austria",
@@ -16,8 +17,16 @@ export enum PowerENUM {
   RUSSIA = "Russia",
 }
 
+
+
 export const ProvTypeSchema = z.nativeEnum(ProvTypeENUM);
-export const PowerSchema = z.nativeEnum(PowerENUM).transform((val) => (val.charAt(0).toUpperCase() + val.slice(1).toLowerCase()));
+export const PowerENUMSchema = z.preprocess((arg) => {
+  if (typeof arg === "string") {
+    // Normalize the string: "austria" or "AUSTRIA" becomes "Austria"
+    return arg.charAt(0).toUpperCase() + arg.slice(1).toLowerCase();
+  }
+  return arg;
+}, z.nativeEnum(PowerENUM));
 
 export const LabelSchema = z.object({
   x: z.number(),
@@ -33,7 +42,8 @@ export const ProvinceSchema = z.object({
   label: LabelSchema,
   type: ProvTypeSchema,
   unit: UnitSchema.optional(),
-  owner: PowerSchema.optional(),
+  owner: PowerENUMSchema.optional(),
+  isSupplyCenter: z.boolean().optional()
 });
 
 export const CoordinateDataSchema = z.object({
@@ -42,4 +52,83 @@ export const CoordinateDataSchema = z.object({
 
 export type Province = z.infer<typeof ProvinceSchema>;
 export type CoordinateData = z.infer<typeof CoordinateDataSchema>;
+enum ProvinceENUM {
+  ANK = "ANK",
+  ARM = "ARM",
+  CON = "CON",
+  MOS = "MOS",
+  SEV = "SEV",
+  STP = "STP",
+  SYR = "SYR",
+  UKR = "UKR",
+  LVN = "LVN",
+  WAR = "WAR",
+  PRU = "PRU",
+  SIL = "SIL",
+  BER = "BER",
+  KIE = "KIE",
+  RUH = "RUH",
+  MUN = "MUN",
+  RUM = "RUM",
+  BUL = "BUL",
+  GRE = "GRE",
+  SMY = "SMY",
+  ALB = "ALB",
+  SER = "SER",
+  BUD = "BUD",
+  GAL = "GAL",
+  VIE = "VIE",
+  BOH = "BOH",
+  TYR = "TYR",
+  TRI = "TRI",
+  FIN = "FIN",
+  SWE = "SWE",
+  NWY = "NWY",
+  DEN = "DEN",
+  HOL = "HOL",
+  BEL = "BEL",
+  SWI = "SWI",
+  VEN = "VEN",
+  PIE = "PIE",
+  TUS = "TUS",
+  ROM = "ROM",
+  APU = "APU",
+  NAP = "NAP",
+  BUR = "BUR",
+  MAR = "MAR",
+  GAS = "GAS",
+  PIC = "PIC",
+  PAR = "PAR",
+  BRE = "BRE",
+  SPA = "SPA",
+  POR = "POR",
+  NAF = "NAF",
+  TUN = "TUN",
+  LON = "LON",
+  WAL = "WAL",
+  LVP = "LVP",
+  YOR = "YOR",
+  EDI = "EDI",
+  CLY = "CLY",
+  NAT = "NAT",
+  NRG = "NRG",
+  BAR = "BAR",
+  BOT = "BOT",
+  BAL = "BAL",
+  SKA = "SKA",
+  HEL = "HEL",
+  NTH = "NTH",
+  ENG = "ENG",
+  IRI = "IRI",
+  MID = "MID",
+  WES = "WES",
+  GOL = "GOL",
+  TYN = "TYN",
+  ADR = "ADR",
+  ION = "ION",
+  AEG = "AEG",
+  EAS = "EAS",
+  BLA = "BLA",
+}
 
+export const ProvinceENUMSchema = z.nativeEnum(ProvinceENUM)
